platform/x86: asus-wmi: Refactor Ally suspend/resume
The CSEE method from ACPI is now called only on module load. This fixes
an issue with USB device loss on Ally 1 after reboot.

Both Ally 1 and Ally X now rely only of the _DSM screen off/on calls
exposed in AMD s2idle to call the CSEE (which is done in the _DSM), plus
a small additional delay in an attempt to allow the MCU time to finish
its internal tasks.

Overall the way this is done is supremely unreliable because it ends up
being a race between MCU, USB subsystem, and power-off, especially if
the MCU powersave mode is enabled and so every Ally 1 and Ally X user
should update to the latest firmware:

Ally-1: v319+
Ally-X: v313+

As of 23/09/24 the FW update release is pending.

Signed-off-by: Luke D. Jones <luke@ljones.dev>