asus_wmi: Add fan_curve interface for asus_armoury use
Due to the PPT settings needing to have fan_curve enabled on at least the
CPU fan to apply them, asus_armoury introduces another attribute
`ppt_enabled` which calls back in to asus_wmi to enable fan curves.

This also needs to be birectional so that both drivers can notify both of
the related attributes: ppt_enabled, and pw<N>_enabled. This is also
expected to be part of future work for custom platform_profile support.

Signed-off-by: Luke Jones <luke@ljones.dev>